removed utils from server module. using the sdk's hera client in dashboard rather than redefining it in src
