+++ sdk/bettmensch_ai/pipelines/__init__.py
@@ -1,3 +1,4 @@
+from bettmensch_ai.client import hera_client  # noqa: F401
 from bettmensch_ai.pipelines.pipeline import (  # noqa: F401
     Pipeline,
     delete_registered_pipeline,
